[Refactor] `common` 모듈 리팩토링 - 불필요한 파일 삭제 - `exception` 패키지를 `api` 패키지 하위로 이동
