--- common/src/main/kotlin/kr/spot/common/domain/events/MemberWithdrawnEvent.kt
@@ -1,5 +1,0 @@
-package kr.spot.common.domain.events
-
-data class MemberWithdrawnEvent(
-    val memberId: Long
-)
