+++ core/src/main/kotlin/kr/spot/core/post/application/PostCommandService.kt
@@ -1,7 +1,7 @@
 package kr.spot.core.post.application
 
 import kr.spot.common.api.status.ErrorStatus
-import kr.spot.common.exception.GeneralException
+import kr.spot.common.api.exception.GeneralException
 import kr.spot.common.id.IdGenerator
 import kr.spot.core.member.application.MemberQueryService
 import kr.spot.core.post.domain.Post
