+++ core/src/main/kotlin/kr/spot/core/post/domain/vo/WriterInfo.kt
@@ -2,7 +2,7 @@ package kr.spot.core.post.domain.vo
 
 import jakarta.persistence.Embeddable
 import kr.spot.common.api.status.ErrorStatus
-import kr.spot.common.exception.GeneralException
+import kr.spot.common.api.exception.GeneralException
 
 @Embeddable
 class WriterInfo(
