+++ core/src/main/kotlin/kr/spot/core/post/infrastructure/jpa/CommentRepository.kt
@@ -1,7 +1,7 @@
 package kr.spot.core.post.infrastructure.jpa
 
 import kr.spot.common.api.status.ErrorStatus
-import kr.spot.common.exception.GeneralException
+import kr.spot.common.api.exception.GeneralException
 import kr.spot.core.post.domain.Comment
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
