+++ study/src/main/kotlin/kr/spot/study/core/application/command/ApplyStudyService.kt
@@ -1,7 +1,7 @@
 package kr.spot.study.core.application.command
 
 import kr.spot.common.api.status.ErrorStatus
-import kr.spot.common.exception.GeneralException
+import kr.spot.common.api.exception.GeneralException
 import kr.spot.common.id.IdGenerator
 import kr.spot.study.core.domain.enums.Decision
 import kr.spot.study.core.domain.enums.StudyMemberStatus
