+++ study/src/main/kotlin/kr/spot/study/core/domain/StudyMember.kt
@@ -6,7 +6,7 @@ import jakarta.persistence.Enumerated
 import jakarta.persistence.Id
 import jakarta.persistence.Table
 import kr.spot.common.api.status.ErrorStatus
-import kr.spot.common.exception.GeneralException
+import kr.spot.common.api.exception.GeneralException
 import kr.spot.study.core.domain.enums.Decision
 import kr.spot.study.core.domain.enums.StudyMemberStatus
 import kr.spot.study.global.domain.BaseEntity
