+++ study/src/main/kotlin/kr/spot/study/core/domain/enums/StudyMemberStatus.kt
@@ -1,7 +1,7 @@
 package kr.spot.study.core.domain.enums
 
 import kr.spot.common.api.status.ErrorStatus
-import kr.spot.common.exception.GeneralException
+import kr.spot.common.api.exception.GeneralException
 
 enum class StudyMemberStatus {
     OWNER, // 스터디 장
