+++ study/src/main/kotlin/kr/spot/study/review/domain/vo/WriterInfo.kt
@@ -3,7 +3,7 @@ package kr.spot.study.review.domain.vo
 import jakarta.persistence.Column
 import jakarta.persistence.Embeddable
 import kr.spot.common.api.status.ErrorStatus
-import kr.spot.common.exception.GeneralException
+import kr.spot.common.api.exception.GeneralException
 
 @Embeddable
 data class WriterInfo(
