+++ study/src/main/kotlin/kr/spot/study/todo/domain/Todo.kt
@@ -3,7 +3,7 @@ package kr.spot.study.todo.domain
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import kr.spot.common.api.status.ErrorStatus
-import kr.spot.common.exception.GeneralException
+import kr.spot.common.api.exception.GeneralException
 import kr.spot.study.global.domain.BaseEntity
 import org.hibernate.annotations.SQLDelete
 import org.hibernate.annotations.SQLRestriction
