+++ worker/src/main/kotlin/kr/spot/worker/WorkerApplication.kt
@@ -3,7 +3,7 @@ package kr.spot.worker
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = ["kr.spot.worker", "kr.spot.common"])
 class WorkerApplication
 
 fun main(args: Array<String>) {
